Optimize seam calculation Batching, seam-mask tracing, cost computation, and higher-precision energy calculations

+++ seamcarver/methods/laplacian.py
@@ -13,6 +13,7 @@ from scipy.ndimage import laplace
 # Import project-specific packages
 from .interface import EnergyMethod
 
+
 class LaplacianEnergy(EnergyMethod):
     """Laplacian energy method for seam carving.
     
@@ -23,8 +24,8 @@ class LaplacianEnergy(EnergyMethod):
     def __call__(self, image) -> np.ndarray:
         """Compute the energy map of the image using the Laplacian operator."""
         # Convert the image to grayscale, then apply the Laplacian operator
-        grayscale_image = np.mean(image, axis=2).astype(np.float16)
+        grayscale_image = np.mean(image, axis=2).astype(np.float32)
         laplacian_image = laplace(grayscale_image, mode='constant', cval=255)
-        energy_tbl = np.abs(laplacian_image).astype(np.float16)
+        energy_tbl = np.abs(laplacian_image)
         
         return energy_tbl # Return the computed energy table
